+++ packages/noka/src/loaders/ControllerLoader/index.ts
@@ -1,4 +1,5 @@
 export * from "./ControllerMeta";
+export * from "./ControllerResult";
 export * from "./RouteMeta";
 export * from "./ContextMeta";
 export * from "./ControllerLoader";